fixup! use plain cjs
