+++ packages/sdk-metrics/test/performance/benchmark/index.js
@@ -1,0 +1,6 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+require('./counter-add');
